requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.5
django-admin-honeypot==1.1.0
django-admin-interface==0.16.3
django-appconf==1.0.4
django-apptemplates==1.5
django-colorfield==0.4.2
django-compressor==2.4.1
django-cors-headers==3.7.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-htmlmin==0.11.0
django-location-field==2.1.0
html5lib==1.1
Pillow==8.3.1
pkg_resources==0.0.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
webencodings==0.5.1
